Updated Controllers and WhiteLabel
Updated all controllers to check if the user is logged in;
If the user is not logged in and attempting to do an action an anonymous user should not be able to do, it will redirect them to the login page.
If the user is attempting to commit an action that is not allowed to them, they are redirected to the uh-oh page that has a button to bring them back home.
Created uh-oh page