Updated Controllers and Service
Updated all controllers to grab User Profile Image URL
CRUD: Created UserService method that pulls current user profile image. If they did not assign one upon registering, the image will be blank as of right now because there is a blank string in the images table.
If for some reason a user was created and row does not exist in the images table for that user, an error will appear in the terminal and assign a default image to that user.